extract DvrpModeRoutingNetworkModule from DvrpRoutingNetworkProvider
