+++ contribs/drt/src/main/java/org/matsim/contrib/drt/run/DrtConfigGroup.java
@@ -35,7 +35,7 @@ import org.apache.log4j.Logger;
 import org.matsim.api.core.v01.TransportMode;
 import org.matsim.contrib.drt.optimizer.insertion.ParallelPathDataProvider;
 import org.matsim.contrib.drt.optimizer.rebalancing.mincostflow.MinCostFlowRebalancingParams;
-import org.matsim.contrib.dvrp.router.DvrpRoutingNetworkProvider;
+import org.matsim.contrib.dvrp.router.DvrpModeRoutingNetworkModule;
 import org.matsim.contrib.dvrp.run.Modal;
 import org.matsim.core.config.Config;
 import org.matsim.core.config.ConfigGroup;
@@ -249,8 +249,7 @@ public final class DrtConfigGroup extends ReflectiveConfigGroup implements Modal
 				"More then one rebalancing parameter sets is specified");
 
 		if (useModeFilteredSubnetwork) {
-			DvrpRoutingNetworkProvider.
-					checkUseModeFilteredSubnetworkAllowed(config, mode);
+			DvrpModeRoutingNetworkModule.checkUseModeFilteredSubnetworkAllowed(config, mode);
 		}
 	}
